utils: move constructor out of header
